GH-696 deleted fixmes and todos that were unrelevant, extracted into files
