+++ MultisigWallet/MultisigWalletApplication/WalletApplicationServiceTests/WalletApplicationServiceTests.swift
@@ -72,7 +72,6 @@ class WalletApplicationServiceTests: BaseWalletApplicationServiceTests {
         XCTAssertEqual(service.ownerAddress(of: .browserExtension), Address.testAccount1.value)
     }
 
-    // FIXME: move to domain model
     func test_whenWalletIsReady_thenHasReadyState() throws {
         createPortfolio()
         service.createNewDraftWallet()
