+++ SafeUIKit/SafeAppUI/AppFlow/MainFlow/Menu/ReplaceBrowserExtension/Intro/RBEIntroViewController.swift
@@ -167,15 +167,8 @@ public class RBEIntroViewController: UIViewController {
         state.retry(controller: self)
     }
 
-    // TODO: duplicated with showTransactionFeeInfo
     @objc public func showNetworkFeeInfo() {
-        let alert = UIAlertController(title: LocalizedString("transaction_fee",
-                                                             comment: "Network fee"),
-                                      message: LocalizedString("transaction_fee_explanation",
-                                                               comment: "Explanatory message"),
-                                      preferredStyle: .alert)
-        alert.addAction(UIAlertAction(title: LocalizedString("close", comment: "Close"), style: .default))
-        present(alert, animated: true)
+        present(TransactionFeeAlertController.create(), animated: true, completion: nil)
     }
 
 }
