+++ SafeUIKit/SafeAppUI/OnboardingFlow/SetupSafeFlow/RecoverSafe/ReviewRecoveryTransactionViewController.swift
@@ -187,15 +187,10 @@ public class ReviewRecoveryTransactionViewController: UIViewController {
         delegate?.reviewRecoveryTransactionViewControllerDidSubmit()
     }
 
-    // TODO: remove duplication
     @objc func showTransactionFeeInfo() {
-        let alert = UIAlertController(title: LocalizedString("transaction_fee", comment: "Network fee"),
-                                      message: LocalizedString("transaction_fee_explanation",
-                                                               comment: "Explanatory message"),
-                                      preferredStyle: .alert)
-        alert.addAction(UIAlertAction(title: LocalizedString("close", comment: "Close"), style: .default))
-        present(alert, animated: true, completion: nil)
+        present(TransactionFeeAlertController.create(), animated: true, completion: nil)
     }
+
 }
 
 extension ReviewRecoveryTransactionViewController: EventSubscriber {
