+++ SafeUIKit/SafeUIKit/Components/QRCodeInput/QRCodeInputTests.swift
@@ -29,21 +29,6 @@ class QRCodeInputTests: XCTestCase {
         XCTAssertNil(delegate.presentedController)
     }
 
-    // TODO: re-enable on Jenkins
-//    func test_textFieldShouldBeginEditing_whenScanOnly_thenCallsDelegateAndReturnsFalse() {
-//        barcodeTextField.editingMode = .scanOnly
-//        XCTAssertFalse(barcodeTextField.textFieldShouldBeginEditing(barcodeTextField))
-//        delay()
-//        assertScannerPresented()
-//    }
-//
-//    func test_openBarcodeSacenner_callsDelegate() {
-//        let button = barcodeTextField.rightView as! UIButton
-//        button.sendActions(for: .touchUpInside)
-//        delay()
-//        assertScannerPresented()
-//    }
-
     func test_didScan_whenCodeIsNotValid_thenDoesNotModifyInput() {
         barcodeTextField.text = "some input"
         barcodeTextField.scanValidatedConverter = { _ in
